+++ inst/examples/divicons_html_app.R
@@ -2,8 +2,8 @@ library(sf)
 library(shiny)
 library(leaflet)
 library(leaflet.extras2)
-# options("shiny.autoreload" = TRUE)
 
+# Sample data
 df <- sf::st_as_sf(atlStorms2005)
 df <- suppressWarnings(st_cast(df, "POINT"))
 df <- df[sample(1:nrow(df), 50, replace = F),]
@@ -15,22 +15,48 @@ ui <- fluidPage(
   ## CSS-style ############
   tags$head(tags$style("
     .globalclass {
-      width: inherit !important;
-      height: inherit !important;
-      font-size: 10px;
-      border-radius: 90%;
-      padding: 4px;
+      width: 80px !important;
+      height: 80px !important;
+      margin-top: -40px !important;
+      margin-left: -40px !important;
+      font-size: 12px;
+      text-align: center;
+      border-radius: 50%;
+      color: black;
+      padding: 5px;
+      box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
+      background-size: cover;
+      background-repeat: no-repeat;
+      background-position: center center;
     }
     .myclass1 {
-      background-color: yellow;
+      background-color: #FF5733;
     }
     .myclass2 {
-      background-color: orange;
+      background-color: #33FF57;
     }
     .myclass3 {
-      background-color: green;
+      background-color: #3357FF;
     }
-    ")),
+    .custom-html {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      flex-direction: column;
+    }
+    .custom-html img {
+      border-radius: 50%;
+      width: 20px;
+      height: 20px;
+      margin-bottom: 5px;
+    }
+    .custom-html .title {
+      font-weight: bold;
+    }
+    .custom-html .subtitle {
+      font-size: 10px;
+    }
+  ")),
   ## CSS-style END ############
   leafletOutput("map", height = 600),
   splitLayout(cellWidths = paste0(rep(20,4), "%"),
@@ -43,14 +69,30 @@ ui <- fluidPage(
 
 ## SERVER ##################
 server <- function(input, output, session) {
+  # Function to get image URL based on class
+  getImageUrls <- function(classes) {
+    urls <- c(
+      "myclass1" = "https://cdn-icons-png.flaticon.com/512/1355/1355883.png",
+      "myclass2" = "https://cdn-icons-png.flaticon.com/512/1356/1356623.png",
+      "myclass3" = "https://cdn-icons-png.flaticon.com/512/1357/1357674.png"
+    )
+    return(urls[classes])
+  }
+
   output$map <- renderLeaflet({
     leaflet()  %>%
       addTiles() %>%
-      addMarkers(data = df, group="normalmarker",
-                 clusterId = "someclusterid2",
-                 clusterOptions = markerClusterOptions()) %>%
+      # addMarkers(data = df, group="normalmarker",
+      #            clusterId = "someclusterid2",
+      #            clusterOptions = markerClusterOptions()) %>%
       addDivicon(data = df
-                 , html = ~Name
+                 , html = ~paste0(
+                   '<div class="custom-html">',
+                   '<img src="', getImageUrls(classes), '">',
+                   '<div class="title">', Name, '</div>',
+                   '<div class="subtitle">MaxWind: ', MaxWind, '</div>',
+                   '</div>'
+                 )
                  , className = ~paste("globalclass", classes)
                  , label = ~Name
                  , layerId = ~ID
@@ -60,12 +102,10 @@ server <- function(input, output, session) {
                               "MaxWind:", MaxWind, "<br>",
                               "MinPress:", MinPress)
                  , options = markerOptions(draggable = TRUE)
-                 , divOptions = list(
-                                     popupAnchor = c(10, 0),
-                                     iconSize = 10)
                  # , clusterId = "someclusterid"
                  # , clusterOptions = markerClusterOptions()
       ) %>%
+      addLabelgun("Divicons", 1) %>%
       hideGroup("normalmarker") %>%
       addLayersControl(overlayGroups = c("Divicons","normalmarker"))
   })
@@ -74,4 +114,5 @@ server <- function(input, output, session) {
   output$mouseout <- renderPrint({input$map_marker_mouseout})
   output$dragend <- renderPrint({input$map_marker_dragend})
 }
+
 shinyApp(ui, server)