+++ inst/htmlwidgets/lfx-divicon/lfx-divicon-bindings.js
@@ -1,120 +1,107 @@
 /* global LeafletWidget, $, L */
 LeafletWidget.methods.addDivicon = function(
-    lats, lngs, icon, layerId, group, options,
-    classes, htmls,
-    popups, popupOptions, labels, labelOptions,
-    clusterId, clusterOptions, divOptions) {
+    lat, lng, layerId, group, options,
+    className, html,
+    popup, popupOptions, label, labelOptions,
+    clusterId, clusterOptions, divOptions,
+    crosstalkOptions) {
 
-    var map = this;
-    console.log("addDivicon")
+    (function() {
+      // Make a Dataframe
+      let df = new LeafletWidget.DataFrame()
+        .col("lat", lat)
+        .col("lng", lng)
+        .col("layerId", layerId)
+        .col("group", group)
+        .col("popup", popup)
+        .col("popupOptions", popupOptions)
+        .col("label", label)
+        .col("labelOptions", labelOptions)
+        .col("className", className)
+        .col("html", html)
+        .cbind(options)
+        .cbind(crosstalkOptions || {});
 
-    // Convert inputs to arrays if they are single strings
-    classes = toArray(classes, lats.length);
-    htmls = toArray(htmls, lats.length);
-    popups = toArray(popups, lats.length);
-    labels = toArray(labels, lats.length);
-    layerIds = toArray(layerId, lats.length);
-
-    for (var i = 0; i < lats.length; i++) {
-      var lat = lats[i];
-      var lng = lngs[i];
-      var iconClass = classes[i];
-      var html = htmls[i];
-      var popupContent = popups[i];
-      var labelContent = labels[i];
-      var layerId = layerIds[i];
-
-      // Create a new marker with DivIcon
-      var divIconOptions = Object.assign({}, divOptions, {
-          className: iconClass,
-          html: html
-      });
-     var divmarker = new L.Marker([lat, lng],
-       Object.assign({}, options, {
-            icon: new L.DivIcon(divIconOptions)
-        }));
-
-      // Bind popup to the marker if popup content is provided
-      if (popupContent) {
-          divmarker.bindPopup(popupContent, popupOptions);
+      // Add Cluster
+      let clusterGroup = this.layerManager.getLayer("cluster", clusterId),
+        cluster = clusterOptions !== null;
+      if (cluster && !clusterGroup) {
+        clusterGroup = L.markerClusterGroup.layerSupport(clusterOptions);
+        if(clusterOptions.freezeAtZoom) {
+          let freezeAtZoom = clusterOptions.freezeAtZoom;
+          delete clusterOptions.freezeAtZoom;
+          clusterGroup.freezeAtZoom(freezeAtZoom);
+        }
+        clusterGroup.clusterLayerStore = new LeafletWidget.ClusterLayerStore(clusterGroup);
       }
+      let extraInfo = cluster ? { clusterId: clusterId } : {};
 
-      // Assign label (tooltip) to marker if label content is provided
-      if (labelContent) {
-          divmarker.bindTooltip(labelContent, labelOptions);
+      for (let i = 0; i < df.nrow(); i++) {
+        if($.isNumeric(df.get(i, "lat")) && $.isNumeric(df.get(i, "lng"))) {
+          (function() {
+
+            let thisId = df.get(i, "layerId");
+            let thisGroup = cluster ? null : df.get(i, "group");
+
+            // Create a new marker with DivIcon
+            var divIconOptions = Object.assign({}, divOptions, {
+                className: df.get(i, "className"),
+                html: df.get(i, "html")
+            });
+            var divmarker = new L.Marker(
+              [df.get(i, "lat"), df.get(i, "lng")],
+              Object.assign({}, options, {
+                  icon: new L.DivIcon(divIconOptions)
+              }));
+
+            if (cluster) {
+              clusterGroup.clusterLayerStore.add(divmarker, thisId);
+            } else {
+              this.layerManager.addLayer(divmarker, "marker", thisId, thisGroup, df.get(i, "ctGroup", true), df.get(i, "ctKey", true));
+            }
+
+            // Bind popup to the marker if popup content is provided
+            let popup = df.get(i, "popup");
+            let popupOptions = df.get(i, "popupOptions");
+            if (popup !== null) {
+              if (popupOptions !== null){
+                divmarker.bindPopup(popup, popupOptions);
+              } else {
+                divmarker.bindPopup(popup);
+              }
+            }
+
+            // Assign label (tooltip) to marker if label content is provided
+            let label = df.get(i, "label");
+            let labelOptions = df.get(i, "labelOptions");
+            if (label !== null) {
+              if (labelOptions !== null) {
+                if(labelOptions.permanent) {
+                  divmarker.bindTooltip(label, labelOptions).openTooltip();
+                } else {
+                  divmarker.bindTooltip(label, labelOptions);
+                }
+              } else {
+                divmarker.bindTooltip(label);
+              }
+            }
+
+            // Add the marker to the map's layer manager
+            this.layerManager.addLayer(divmarker, "marker", thisId, thisGroup);
+
+            // Add Listener
+            divmarker.on("click", LeafletWidget.methods.mouseHandler(this.id, thisId, thisGroup, "marker_click", extraInfo), this);
+            divmarker.on("mouseover", LeafletWidget.methods.mouseHandler(this.id, thisId, thisGroup, "marker_mouseover", extraInfo), this);
+            divmarker.on("mouseout", LeafletWidget.methods.mouseHandler(this.id, thisId, thisGroup, "marker_mouseout", extraInfo), this);
+            divmarker.on("dragend", LeafletWidget.methods.mouseHandler(this.id, thisId, thisGroup, "marker_dragend", extraInfo), this);
+
+          }).call(this);
+        }
       }
-
-      // Add the marker to the map's layer manager
-      map.layerManager.addLayer(divmarker, "marker", layerId, group);
-
-      divmarker.on("click", LeafletWidget.methods.mouseHandler(map.id, layerId, group, "marker_click", ""), this);
-      divmarker.on("mouseover", LeafletWidget.methods.mouseHandler(map.id, layerId, group, "marker_mouseover", ""), this);
-      divmarker.on("mouseout", LeafletWidget.methods.mouseHandler(map.id, layerId, group, "marker_mouseout", ""), this);
-      divmarker.on("dragend", LeafletWidget.methods.mouseHandler(map.id, layerId, group, "marker_dragend", ""), this);
-
-      if (clusterId) {
-        map.layerManager.addLayer(divmarker, "cluster", clusterId, group);
+      if (cluster) {
+        this.layerManager.addLayer(clusterGroup, "cluster", clusterId, group);
       }
+    }).call(this);
 
-    }
 };
 
-
-// Convert single string inputs to arrays
-function toArray(input, length) {
-    if (typeof input === 'string') {
-        return Array(length).fill(input);
-    }
-    return input;
-}
-
-
-/*
-addDivicon(map, df, group, markerFunc) {
-  (function() {
-    for (let i = 0; i < df.nrow(); i++) {
-      if($.isNumeric(df.get(i, "lat")) && $.isNumeric(df.get(i, "lng"))) {
-        (function() {
-          let marker = markerFunc(df, i);
-          let thisId = df.get(i, "layerId");
-          let thisGroup = cluster ? null : df.get(i, "group");
-          this.layerManager.addLayer(marker, "marker", thisId, thisGroup, df.get(i, "ctGroup", true), df.get(i, "ctKey", true));
-
-          let popup = df.get(i, "popup");
-          let popupOptions = df.get(i, "popupOptions");
-          if (popup !== null) {
-            if (popupOptions !== null){
-              marker.bindPopup(popup, popupOptions);
-            } else {
-              marker.bindPopup(popup);
-            }
-          }
-          let label = df.get(i, "label");
-          let labelOptions = df.get(i, "labelOptions");
-          if (label !== null) {
-            if (labelOptions !== null) {
-              if(labelOptions.permanent) {
-                marker.bindTooltip(label, labelOptions).openTooltip();
-              } else {
-                marker.bindTooltip(label, labelOptions);
-              }
-            } else {
-              marker.bindTooltip(label);
-            }
-          }
-          //marker.on("click", mouseHandler(this.id, thisId, thisGroup, "marker_click"), this);
-          //marker.on("mouseover", mouseHandler(this.id, thisId, thisGroup, "marker_mouseover"), this);
-          //marker.on("mouseout", mouseHandler(this.id, thisId, thisGroup, "marker_mouseout"), this);
-          //marker.on("dragend", mouseHandler(this.id, thisId, thisGroup, "marker_dragend"), this);
-
-        }).call(this);
-      }
-    }
-
-    if (cluster) {
-      this.layerManager.addLayer(clusterGroup, "cluster", clusterId, group);
-    }
-  }).call(map);
-}
-
-*/